much

+++ scraper/services/ocr.js
@@ -14,23 +14,15 @@ const { getCannabinoidObj, getTerpeneObj } = require('./strings.js');
 const badImages = [];
 
 const configWNCTerpenesTitle = {
-  rectangle: { top: 1397, left: 292, width: 365, height: 380 },
+  rectangle: { top: 1352, left: 284, width: 400, height: 188 },
 }
 
 const configWNCCannabinoidsTitle = {
   rectangle: { top: 2100, left: 301, width: 485, height: 94 },
 }
-
-/*
-'ambigs_debug_level'
-'user_words_suffix'
-'user_patterns_suffix'
-'user_patterns_suffix',
-'load_system_dawg',
- 'load_freq_dawg',
- 'load_unambig_dawg','load_punc_dawg', 'load_number_dawg', 'load_bigram_dawg',
-'tessedit_ocr_engine_mode', 'tessedit_init_config_only', 'language_model_ngram_on', 'language_model_use_sigmoidal_certainty'];
-*/
+const configWNCCannabinoidsTitle2 = {
+  rectangle: { top: 1522, left: 86, width: 621, height: 313 },
+}
 
 const configWNCTerpenes = {
   rectangle: { top: 1722, left: 320, width: 1939, height: 1361 },
@@ -38,6 +30,10 @@ const configWNCTerpenes = {
 
 const configWNCCannabinoids = {
   rectangle: { top: 2511, left: 555, width: 2457, height: 1484 },
+}
+
+const configWNCCannabinoids2 = {
+  rectangle: { top: 1756, left: 108, width: 3181, height: 1556 },
 }
 
 const jpgNameFromUrl = (url) => {
@@ -69,7 +65,7 @@ const getAndProcessJpg = async (url, isDev) => {
     const response = await axios.get(url, { responseType: 'arraybuffer' });
 
     if (response.status !== 200) {
-      console.log(`Failed to download image 1`);
+      // console.log(`Failed to download image 1`);
       badImages.push({ url, error: response.status });
       return null;
     }
@@ -77,7 +73,7 @@ const getAndProcessJpg = async (url, isDev) => {
     const contentType = response.headers['content-type'];
 
     if (contentType !== 'image/jpeg') {
-      console.log(`Invalid content type: ${contentType}`);
+      // console.log(`Invalid content type: ${contentType}`);
       badImages.push({ url, error: contentType });
       return null
     }
@@ -87,7 +83,7 @@ const getAndProcessJpg = async (url, isDev) => {
     const gmResponse = await gm(response.data, jpgName).resize(4000).sharpen(0.5, 0.5).quality(100).setFormat('jpg')
 
     const jpgBuffer = await gmToBuffer(gmResponse);
-    fs.writeFileSync('image.jpg', jpgBuffer);
+
     return jpgBuffer;
 
   }
@@ -96,55 +92,52 @@ const getAndProcessJpg = async (url, isDev) => {
     console.error(`Failed to process image 2`, error);
     badImages.push({ url, error });
 
-    return null;
+    return JSON.stringify(error);
   }
 }
 
 async function recognize(url) {
 
   try {
-    console.log('url', url)
-    if (url.includes('Pineapple') ||
-      url.includes('CBD-Diamond') ||
-      url.includes('BagsGroupSho')) {
-      console.log('pineapple')
-      return null;
 
-    }
+    console.log('URL', url)
 
     const worker = await createWorker("eng", 1, {
       user_patterns_file: './tessdata/eng.user-patterns',
       tessdata: './tessdata',
       userPatterns: './tessdata/eng.user-patterns',
       tessedit_write_images: true,
-
+      errorHandler: (err) => { console.error('Error:', err) },
     });
 
     await worker.loadLanguage('eng');
 
-    await worker.initialize();
+    await worker.initialize('eng');
 
     await worker.setParameters({
       tessedit_pageseg_mode: 4,
-      tessedit_rejection_debug: 1,
     });
 
     const isDev = process.env.NODE_ENV !== 'production';
-    console.log('recognizing', url)
-
 
     const jpgBuffer = await getAndProcessJpg(url, isDev);
+
     if (!jpgBuffer) {
-      console.log('skipping', url);
+      // console.log('skipping empty', url);
       badImages.push(url);
+      fs.appendFileSync('badImages.json', JSON.stringify(badImages, null, 2));
       await worker.terminate();
+
       return null;
     }
-    fs.writeFileSync('image.jpg', jpgBuffer);
+
+    //fs.writeFileSync('image.jpg', jpgBuffer);
 
     const terpenes = [];
 
     const cannabinoids = [];
+
+    const jpgName = jpgNameFromUrl(url);
 
     let title = await worker.recognize(jpgBuffer, configWNCTerpenesTitle);
 
@@ -152,7 +145,7 @@ async function recognize(url) {
 
       console.log('----- terpenes -----');
 
-      const result = await worker.recognize('image.jpg', configWNCTerpenes);
+      const result = await worker.recognize(jpgBuffer, configWNCTerpenes);
 
       const textArray = result.data.text.split('\n');
 
@@ -167,12 +160,13 @@ async function recognize(url) {
       await worker.terminate();
 
       terpenes.sort((a, b) => b.pct - a.pct);
-      return {
-        terpenes
-      }
+
+      return { terpenes }
+
     }
 
-    title = await worker.recognize('image.jpg', configWNCCannabinoidsTitle);
+    await worker.terminate();
+    return null;
 
     title = await worker.recognize(jpgBuffer, configWNCCannabinoidsTitle);
 
@@ -187,21 +181,45 @@ async function recognize(url) {
       for (const text of textArray) {
 
         const line = getCannabinoidObj(text);
-        if (!line) {
-          console.log('not line');
-        }
 
         cannabinoids.push(line);
 
       }
 
+      await worker.terminate();
+
+      cannabinoids.sort((a, b) => b.pct - a.pct);
+
+      return { cannabinoids }
+
     }
 
-    await worker.terminate();
+    title = await worker.recognize(jpgBuffer, configWNCCannabinoidsTitle2);
 
-    cannabinoids.sort((a, b) => b.pct - a.pct);
+    if (title.data.text.toLowerCase().includes('cannabinoids')) {
 
-    return { terpenes, cannabinoids }
+      console.log('----- cannabinoids 2 -----');
+
+      const result = await worker.recognize(jpgBuffer, configWNCCannabinoids2);
+
+      const textArray = result.data.text.split('\n');
+
+      for (const text of textArray) {
+
+        const line = getCannabinoidObj(text);
+
+        cannabinoids.push(line);
+
+      }
+
+      await worker.terminate();
+
+      cannabinoids.sort((a, b) => b.pct - a.pct);
+
+      return { cannabinoids }
+
+    }
+
 
   }
   catch (error) {
@@ -210,11 +228,28 @@ async function recognize(url) {
     }
     console.error(`Failed to recognize image: ${error} `);
     fs.writeFileSync('error.txt', JSON.stringify(error, null, 2));
-    return null;
+
+    return error;
   }
 
 
 }
+
+const recognizeWithSave = async (imagePath, options) => {
+  // Process the image with Tesseract.js and save the optimized image to disk
+  const worker = await createWorker();
+  await worker.loadLanguage('eng');
+  await worker.initialize('eng');
+  await worker.setParameters(options);
+  const { data, hocr } = await worker.recognize(imagePath);
+  const optimizedImagePath = `${imagePath}.optimized.jpg`;
+  const optimizedImageBuffer = await worker.getBuffer('jpg');
+  fs.writeFileSync(optimizedImagePath, optimizedImageBuffer);
+  await worker.terminate();
+
+  // Return the OCR result and the path to the optimized image
+  return { data, hocr, optimizedImagePath };
+};
 /*
 (async () => {
   const result = await recognize('http://localhost:5173/BUFFER.jpg');
@@ -222,5 +257,6 @@ async function recognize(url) {
 })();
 */
 module.exports = {
-  recognize
+  recognize,
+  recognizeWithSave
 };
